Added SOLR_select in event cache_retry is not available
