+++ exhibits/templates/exhibits/exhibit_items.html
@@ -1,11 +1,11 @@
 <div class="row primarysource">
   {% for item in exhibitItems %}
+    {% with item.solrData as exhibitItemData %}
     <div class="col-xs-6 col-sm-3 col-md-2">
-        <a class="primarysource__link js-exhibit-item" href="{% url url_name url_id item.item_id %}" title="{% if item.solrData %}{{ item.solrData.title.0 }}{% else %}{{ item.custom_title }}{% endif %}">
+        <a class="primarysource__link js-exhibit-item" href="{% url url_name url_id item.item_id %}" title="{% if exhibitItemData %}{{ exhibitItemData.title.0 }}{% else %}{{ item.custom_title }}{% endif %}">
 
         <div class="thumbnail__container">
-          {% if item.solrData %}
-            {% with item.solrData as exhibitItemData %}
+          {% if exhibitItemData %}
               {% if exhibitItemData.reference_image_md5 %}
                 <img class="thumbnail__image"
                 src="{{ thumbnailUrl }}crop/210x210/{{ exhibitItemData.reference_image_md5 }}"
@@ -34,7 +34,6 @@
                 <div class="thumbnail__tile-icon fa-picture-o"></div>
                 {% endif %}
               {% endif %}
-            {% endwith %}
           {% elif item.custom_crop %}
             <img class="thumbnail__image" src="{{ thumbnailUrl }}crop/210x210/{{ item.custom_crop }}"
             srcset="{{ thumbnailUrl }}crop/210x210/{{ item.custom_crop }} 60w, {{ thumbnailUrl }}crop/310x310/{{ item.custom_crop }} 50w"
@@ -42,8 +41,8 @@
           {% endif %}
 
           <div class="exhibit__caption">
-            {% if item.solrData %}
-              {{ item.solrData.title.0 }}
+            {% if exhibitItemData %}
+              {{ exhibitItemData.title.0 }}
             {% elif item.custom_title %}
               {{ item.custom_title }}
             {% else %}
@@ -53,7 +52,7 @@
         </div>
         </a>
     </div>
-
+    {% endwith %}
     {% if forloop.counter == 12 %}
       <div class="js-exhibit-items-overflow hidden-content" style="width: 100%;">
     {% endif %}
