+++ exhibits/templates/exhibits/item_content.html
@@ -18,8 +18,8 @@
 
   <div class="exhibit__modal-content-wrapper">
   {% with clipto="500x500" %}
-  {% if exhibitItem.solrData %}
   {% with item=exhibitItem.solrData %}
+  {% if item %}
     <div class="exhibit__modal-content">
       <div class="obj-container__simple-{% if 'reference_image_md5' in item %}image{% else %}tile{% endif %}">
         <a class="obj__link" href="{% if not active_tab %}{% url 'calisphere:itemView' item.id %}{% endif %}" data-pjax="js-pageContent">
@@ -119,7 +119,6 @@
           </dl>
       </div>
     </div>
-  {% endwith %}
   {% else %}
     <div class="exhibit__modal-content">
       <div class="obj-container__simple-{% if exhibitItem.custom_crop %}image{% else %}tile{% endif %}">
@@ -154,6 +153,7 @@
       </div>
     </div>
   {% endif %}
+  {% endwith %}
   {% endwith %}
   </div>
